+++ src/components.js
@@ -1,22 +1,22 @@
 import InertiaForm from './InertiaForm.vue'
-import Checkbox from './components/Checkbox.vue'
+import CheckboxInput from './components/CheckboxInput.vue'
 import DatepickerInput from './components/DatepickerInput.vue'
 import FormInput from './components/FormInput.vue'
 import InputError from './components/InputError.vue'
 import InputLabel from './components/InputLabel.vue'
-import Select from './components/Select.vue'
+import SelectInput from './components/SelectInput.vue'
 import TextareaInput from './components/TextareaInput.vue'
 import TextInput from './components/TextInput.vue'
 import PrimaryButton from './components/PrimaryButton.vue'
 
 export default {
     InertiaForm: InertiaForm,
-    Checkbox,
+    CheckboxInput,
     DatepickerInput,
     FormInput,
     InputError,
     InputLabel,
-    Select,
+    SelectInput,
     TextareaInput,
     TextInput,
     PrimaryButton,