+++ CHANGELOG.md
@@ -6,12 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Index file
+### Changed
+- Rename `Checkbox` to `CheckboxInput` and `Select` to `SelectInput`  
 
 ## 0.1.3 - 2023-09-10
 
 ## 0.1.2 - 2023-09-07
+### Fixed
+- Datepicker format
+- Publish package
 
 ## 0.1.1 - 2023-09-05
+### Fixed
+-  Build before publish package
 
 ## 0.1.0 - 2023-07-02
 ### Added